fashion-recommender/requirements.txt
tensorflow==2.12.0
numpy>=1.21.0
Pillow>=9.0.0
pymongo>=4.0.0
tqdm>=4.65.0
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.26.0
pydantic>=1.8.0
python-multipart>=0.0.5  # Required for FastAPI file uploads 